standalone menu - split into separate file
